Rename HttpError to UnexpectedStatusCode, move to ResponseException object
