+++ core/src/main/scala/sttp/client4/ResponseException.scala
@@ -7,10 +7,10 @@ import sttp.model.ResponseMetadata
   * successfully.
   *
   * A response exception can itself be one of two cases:
-  *   - a [[HttpError]], when the response code is other than 2xx (or whatever is considered "success" by the response
-  *     handling description); the body is deserialized to `HE`
-  *   - a [[DeserializationException]], when there's an error during deserialization (this includes deserialization
-  *     exceptions of both the success and error branches)
+  *   - [[ResponseException.UnexpectedStatusCode]], when the response code is other than 2xx (or whatever is considered
+  *     "success" by the response handling description); the body is deserialized to `HE`
+  *   - [[ResponseException.DeserializationException]], when there's an error during deserialization (this includes
+  *     deserialization exceptions of both the success and error branches)
   *
   * This type is often used as the left-side of a top-level either (where the right-side represents a successful request
   * and deserialization). When thrown/returned when sending a request (e.g. in `...OrFailed` response handling
@@ -26,24 +26,27 @@ sealed abstract class ResponseException[+HE](
     val response: ResponseMetadata
 ) extends Exception(error, cause.orNull)
 
-/** Represents an http error, where the response was received successfully, but the status code is other than the
-  * expected one (typically other than 2xx).
-  *
-  * @tparam HE
-  *   The type of the body to which the error response is deserialized.
-  */
-case class HttpError[+HE](body: HE, override val response: ResponseMetadata)
-    extends ResponseException[HE](s"statusCode: ${response.code}, response: $body", None, response)
-
-/** Represents an error that occurred during deserialization of `body`. */
-case class DeserializationException(body: String, cause: Exception, override val response: ResponseMetadata)
-    extends ResponseException[Nothing](
-      cause.getMessage(),
-      Some(cause),
-      response
-    )
-
 object ResponseException {
+
+  /** Represents an error, where the response was received successfully, but the status code is other than the expected
+    * one (typically other than 2xx).
+    *
+    * @tparam HE
+    *   The type of the body to which the error response is deserialized.
+    */
+  case class UnexpectedStatusCode[+HE](body: HE, override val response: ResponseMetadata)
+      extends ResponseException[HE](s"statusCode: ${response.code}, response: $body", None, response)
+
+  /** Represents an error that occurred during deserialization of `body`. */
+  case class DeserializationException(body: String, cause: Exception, override val response: ResponseMetadata)
+      extends ResponseException[Nothing](
+        cause.getMessage(),
+        Some(cause),
+        response
+      )
+
+  //
+
   @tailrec def find(exception: Throwable): Option[ResponseException[_]] =
     Option(exception) match {
       case Some(e: ResponseException[_]) => Some(e)
