+++ json/json4s/src/main/scala/sttp/client4/json4s/SttpJson4sApi.scala
@@ -6,6 +6,8 @@ import sttp.client4.internal.Utf8
 import sttp.client4.json._
 import sttp.model._
 import sttp.client4.ResponseAs.deserializeEitherOrThrow
+import sttp.client4.ResponseException.DeserializationException
+import sttp.client4.ResponseException.UnexpectedStatusCode
 
 trait SttpJson4sApi {
 
@@ -18,7 +20,8 @@ trait SttpJson4sApi {
 
   /** If the response is successful (2xx), tries to deserialize the body from a string into JSON. Returns:
     *   - `Right(b)` if the parsing was successful
-    *   - `Left(HttpError(String))` if the response code was other than 2xx (deserialization is not attempted)
+    *   - `Left(UnexpectedStatusCode(String))` if the response code was other than 2xx (deserialization is not
+    *     attempted)
     *   - `Left(DeserializationException)` if there's an error during deserialization
     */
   def asJson[B: Manifest](implicit
@@ -49,7 +52,7 @@ trait SttpJson4sApi {
   /** Tries to deserialize the body from a string into JSON, using different deserializers depending on the status code.
     * Returns:
     *   - `Right(B)` if the response was 2xx and parsing was successful
-    *   - `Left(HttpError(E))` if the response was other than 2xx and parsing was successful
+    *   - `Left(UnexpectedStatusCode(E))` if the response was other than 2xx and parsing was successful
     *   - `Left(DeserializationException)` if there's an error during deserialization
     */
   def asJsonEither[E: Manifest, B: Manifest](implicit
@@ -58,8 +61,10 @@ trait SttpJson4sApi {
   ): ResponseAs[Either[ResponseException[E], B]] =
     asJson[B].mapLeft { (l: ResponseException[String]) =>
       l match {
-        case HttpError(e, meta) =>
-          ResponseAs.deserializeCatchingExceptions(deserializeJson[E])(e, meta).fold(identity, HttpError(_, meta))
+        case UnexpectedStatusCode(e, meta) =>
+          ResponseAs
+            .deserializeCatchingExceptions(deserializeJson[E])(e, meta)
+            .fold(identity, UnexpectedStatusCode(_, meta))
         case de @ DeserializationException(_, _, _) => de
       }
     }.showAsJsonEither
