+++ json/play-json/src/main/scala/sttp/client4/playJson/SttpPlayJsonApi.scala
@@ -8,6 +8,8 @@ import sttp.model.MediaType
 
 import scala.util.{Failure, Success, Try}
 import sttp.client4.ResponseAs.deserializeEitherWithErrorOrThrow
+import sttp.client4.ResponseException.UnexpectedStatusCode
+import sttp.client4.ResponseException.DeserializationException
 
 trait SttpPlayJsonApi {
 
@@ -17,7 +19,8 @@ trait SttpPlayJsonApi {
 
   /** If the response is successful (2xx), tries to deserialize the body from a string into JSON. Returns:
     *   - `Right(b)` if the parsing was successful
-    *   - `Left(HttpError(String))` if the response code was other than 2xx (deserialization is not attempted)
+    *   - `Left(UnexpectedStatusCode(String))` if the response code was other than 2xx (deserialization is not
+    *     attempted)
     *   - `Left(DeserializationException)` if there's an error during deserialization
     */
   def asJson[B: Reads: IsOption]: ResponseAs[Either[ResponseException[String], B]] =
@@ -39,14 +42,14 @@ trait SttpPlayJsonApi {
   /** Tries to deserialize the body from a string into JSON, using different deserializers depending on the status code.
     * Returns:
     *   - `Right(B)` if the response was 2xx and parsing was successful
-    *   - `Left(HttpError(E))` if the response was other than 2xx and parsing was successful
+    *   - `Left(UnexpectedStatusCode(E))` if the response was other than 2xx and parsing was successful
     *   - `Left(DeserializationException)` if there's an error during deserialization
     */
   def asJsonEither[E: Reads: IsOption, B: Reads: IsOption]: ResponseAs[Either[ResponseException[E], B]] =
     asJson[B].mapLeft { (l: ResponseException[String]) =>
       l match {
-        case HttpError(e, meta) =>
-          deserializeJson[E].apply(e).fold(DeserializationException(e, _, meta), HttpError(_, meta))
+        case UnexpectedStatusCode(e, meta) =>
+          deserializeJson[E].apply(e).fold(DeserializationException(e, _, meta), UnexpectedStatusCode(_, meta))
         case de: DeserializationException => de
       }
     }.showAsJsonEither
