+++ json/play-json/src/test/scala/sttp/client4/PlayJsonTests.scala
@@ -9,6 +9,7 @@ import sttp.model.StatusCode
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import sttp.client4.json.RunResponseAs
+import sttp.client4.ResponseException.DeserializationException
 
 class PlayJsonTests extends AnyFlatSpec with Matchers with EitherValues {
