+++ json/zio-json/src/main/scalajvm/sttp/client4/ziojson/SttpZioJsonApiExtensions.scala
@@ -2,8 +2,6 @@ package sttp.client4.ziojson
 
 import sttp.capabilities.Effect
 import sttp.capabilities.zio.ZioStreams
-import sttp.client4.DeserializationException
-import sttp.client4.HttpError
 import sttp.client4.ResponseException
 import sttp.client4.StreamResponseAs
 import sttp.client4.asStreamWithMetadata
@@ -11,6 +9,8 @@ import zio.Task
 import zio.ZIO
 import zio.json.JsonDecoder
 import zio.stream.ZPipeline
+import sttp.client4.ResponseException.UnexpectedStatusCode
+import sttp.client4.ResponseException.DeserializationException
 
 trait SttpZioJsonApiExtensions { this: SttpZioJsonApi =>
   def asJsonStream[B: JsonDecoder]
@@ -21,7 +21,7 @@ trait SttpZioJsonApiExtensions { this: SttpZioJsonApi =>
         .map(Right(_))
         .catchSome { case e: Exception => ZIO.left(DeserializationException("", e, meta)) }
     ).mapWithMetadata {
-      case (Left(s), meta) => Left(HttpError(s, meta))
+      case (Left(s), meta) => Left(UnexpectedStatusCode(s, meta))
       case (Right(s), _)   => s
     }
 }
