+++ json/zio1-json/src/main/scalajvm/sttp/client4/ziojson/SttpZioJsonApiExtensions.scala
@@ -2,8 +2,6 @@ package sttp.client4.ziojson
 
 import sttp.capabilities.Effect
 import sttp.capabilities.zio.ZioStreams
-import sttp.client4.DeserializationException
-import sttp.client4.HttpError
 import sttp.client4.ResponseException
 import sttp.client4.StreamResponseAs
 import sttp.client4.asStreamWithMetadata
@@ -12,6 +10,8 @@ import zio.ZIO
 import zio.blocking.Blocking
 import zio.json.JsonDecoder
 import zio.stream.ZTransducer
+import sttp.client4.ResponseException.DeserializationException
+import sttp.client4.ResponseException.UnexpectedStatusCode
 
 trait SttpZioJsonApiExtensions { this: SttpZioJsonApi =>
   def asJsonStream[B: JsonDecoder]
@@ -22,7 +22,7 @@ trait SttpZioJsonApiExtensions { this: SttpZioJsonApi =>
         .map(Right(_))
         .catchSome { case e: Exception => ZIO.left(DeserializationException("", e, meta)) }
     ).mapWithMetadata {
-      case (Left(s), meta) => Left(HttpError(s, meta))
+      case (Left(s), meta) => Left(UnexpectedStatusCode(s, meta))
       case (Right(s), _)   => s
     }
 }
